+++ better-better-booru.user.js
@@ -3,7 +3,7 @@
 // @namespace      https://greasyfork.org/scripts/3575-better-better-booru
 // @author         otani, modified by Jawertae, A Pseudonymous Coder & Moebius Strip.
 // @description    Several changes to make Danbooru much better.
-// @version        8.2
+// @version        8.2.1
 // @updateURL      https://greasyfork.org/scripts/3575-better-better-booru/code/better_better_booru.meta.js
 // @downloadURL    https://greasyfork.org/scripts/3575-better-better-booru/code/better_better_booru.user.js
 // @match          *://*.donmai.us/*
@@ -71,6 +71,23 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		}
 
 		return encoded;
+	};
+
+	Element.prototype.bbbParent = function(parentName, limit) {
+		// Search an element's successive parent nodes for a specific tag name or until the amount limit is hit.
+		var el = this;
+		var search = parentName.toUpperCase();
+
+		for (var i = 0, il = limit || 1; i < il; i++) {
+			var parent = el.parentNode;
+
+			if (parent && parent.tagName === search)
+				return parent;
+			else
+				el = parent;
+		}
+
+		return null;
 	};
 
 	Element.prototype.bbbGetPadding = function() {
@@ -340,7 +357,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			swapped: false // Whether the post content has been changed between the original and sample versions.
 		},
 		options: { // Setting options and data.
-			bbb_version: "8.2",
+			bbb_version: "8.2.1",
 			add_popular_link: newOption("checkbox", false, "Add Popular Link", "Add a link to the popular listing to the \"posts\" submenu"),
 			add_random_post_link: newOption("checkbox", false, "Add Random Link", "Add a link to a random post to the post sidebar options menu."),
 			alternate_image_swap: newOption("checkbox", false, "Alternate Image Swap", "Switch between the sample and original image by clicking the image. <tiphead>Note</tiphead>Notes can be toggled by using the link in the sidebar options section."),
@@ -370,7 +387,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			enable_status_message: newOption("checkbox", true, "Enable Status Message", "When requesting information from Danbooru, display the request status in the lower right corner."),
 			endless_default: newOption("dropdown", "disabled", "Default", "Enable endless pages on the favorites, search, pool, and favorite group listings. <tipdesc>Off:</tipdesc> Start up with all features off. <tipdesc>On:</tipdesc> Start up with all features on.<tipdesc>Paused:</tipdesc> Start up with all features on, but do not append new pages until the \"load more\" button is clicked. <tiphead>Note</tiphead>When not set to disabled, endless pages can be toggled between off and on/paused by using the \"E\" hotkey or the \"endless\" link next to the \"listing\" link in the page submenu. <tiphead>Tip</tiphead>The \"new tab/window\" and \"fixed paginator\" options can provide additional customization for endless pages.", {txtOptions:["Disabled:disabled", "Off:off", "On:on", "Paused:paused"]}),
 			endless_fill: newOption("checkbox", false, "Fill Pages", "When appending pages with missing thumbnails caused by hidden posts or removed duplicate posts, retrieve thumbnails from the following pages and add them to the new page until the desired number of thumbnails is reached. <tiphead>Note</tiphead>If using page separators, the displayed page number for appended pages composed of thumbnails from multiple Danbooru pages will be replaced by a range consisting of the first and last pages from which thumbnails were retrieved."),
-			endless_pause_interval: newOption("dropdown", 0, "Pause Interval", "Pause endless pages each time the number of pages reaches a multiple of the selected amount.", {txtOptions:["Disabled:0"], numRange:[1,100]}),
+			endless_pause_interval: newOption("dropdown", 0, "Pause Interval", "Pause endless pages each time the number of pages reaches a multiple of the selected amount. <tiphead>Note</tiphead> When this option is enabled, the \"Shift + E\" hotkey can be used to continue loading more pages after pausing.", {txtOptions:["Disabled:0"], numRange:[1,100]}),
 			endless_preload: newOption("checkbox", false, "Preload Next Page", "Start loading the next page as soon as possible.<tiphead>Note</tiphead>A preloaded page will not be appended until the scroll limit is reached."),
 			endless_remove_dup: newOption("checkbox", false, "Remove Duplicates", "When appending new pages, remove posts that already exist in the listing from the new page.<tiphead>Note</tiphead>Duplicate posts are caused by the addition of new posts to the beginning of a listing or changes to the order of the posts."),
 			endless_scroll_limit: newOption("dropdown", 500, "Scroll Limit", "Set the minimum amount of pixels that the window can have left to vertically scroll before it starts appending the next page.", {numList:[0,50,100,150,200,250,300,350,400,450,500,550,600,650,700,750,800,850,900,950,1000,1050,1100,1150,1200,1250,1300,1350,1400,1450,1500]}),
@@ -402,7 +419,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			post_resize_mode: newOption("dropdown", "width", "Resize Mode", "Choose how to shrink large post content to fit the browser window when initially loading a post.", {txtOptions:["Width:width", "Height:height", "Width & Height:all"]}),
 			post_tag_scrollbars: newOption("dropdown", 0, "Post Tag Scrollbars", "Limit the length of the sidebar tag lists for posts by restricting them to a set height in pixels. For lists that exceed the set height, a scrollbar will be added to allow the rest of the list to be viewed.<tiphead>Note</tiphead>When using \"remove tag headers\", this option will limit the overall length of the combined list.", {txtOptions:["Disabled:0"], numList:[50,100,150,200,250,300,350,400,450,500,550,600,650,700,750,800,850,900,950,1000,1050,1100,1150,1200,1250,1300,1350,1400,1450,1500]}),
 			post_tag_titles: newOption("checkbox", false, "Post Tag Titles", "Change the page titles for posts to a full list of the post tags."),
-			quick_search: newOption("dropdown", "disabled", "Quick Search", "Add a new search box to the upper right corner of the window viewport that allows searching through the current thumbnails for specific posts. <tipdesc>Fade:</tipdesc> Fade all posts that don't match in the thumbnail listing. <tipdesc>Remove:</tipdesc> Remove all posts that don't match from the thumbnail listing. <tiphead>Directions</tiphead>Please read the \"thumbnail matching rules\" section under the help tab for information about creating searches. <br><br>The search starts minimized in the upper right corner. Left clicking the main icon will open and close the search. Right clicking the main icon will completely reset the search. Holding down shift while left clicking the main icon will toggle an active search's pinned status. <br><br>While open, the search can be entered/updated in the search box and the pinned status can be toggled by clicking the pushpin icon. Clicking the negative icon next to the pushpin icon will tell quick search to negate/invert the search being submitted. If no changes are made to an active search, submitting it a second time will reset the quick search. <tiphead>Notes</tiphead>Options labeled with \"pinned\" will make searches default to being pinned. <br><br>A pinned search will persist across other pages in the same browsing session for that tab until it ends or the search is unpinned. <br><br>When not set to disabled, the quick search can be opened by using the \"F\" hotkey. Additionally, an active search can be reset by using \"Shift + F\". Pressing \"Escape\" while the quick search is open will close it.", {txtOptions:["Disabled:disabled", "Fade:fade", "Fade (Pinned):fade pinned", "Remove:remove", "Remove (Pinned):remove pinned"]}),
+			quick_search: newOption("dropdown", "disabled", "Quick Search", "Add a new search box to the upper right corner of the window viewport that allows searching through the current thumbnails for specific posts. <tipdesc>Fade:</tipdesc> Fade all posts that don't match in the thumbnail listing. <tipdesc>Remove:</tipdesc> Remove all posts that don't match from the thumbnail listing. <tiphead>Directions</tiphead>Please read the \"thumbnail matching rules\" section under the help tab for information about creating searches. <br><br>The search starts minimized in the upper right corner. Left clicking the main icon will open and close the search. Right clicking the main icon will completely reset the search. Holding down shift while left clicking the main icon will toggle an active search's pinned status. Holding down control while left clicking the main icon will toggle an active search's negated status.<br><br>While open, the search can be entered/updated in the search box and the pinned status can be toggled by clicking the pushpin icon. Clicking the negative icon next to the pushpin icon will tell quick search to negate/invert the search being submitted. If no changes are made to an active search, submitting it a second time will reset the quick search. <tiphead>Notes</tiphead>Options labeled with \"pinned\" will make searches default to being pinned. <br><br>A pinned search will persist across other pages in the same browsing session for that tab until it ends or the search is unpinned. <br><br>When not set to disabled, the quick search can be opened by using the \"F\" hotkey. Additionally, an active search can be reset by using \"Shift + F\". Pressing \"Escape\" while the quick search is open will close it.", {txtOptions:["Disabled:disabled", "Fade:fade", "Fade (Pinned):fade pinned", "Remove:remove", "Remove (Pinned):remove pinned"]}),
 			remove_tag_headers: newOption("checkbox", false, "Remove Tag Headers", "Remove the \"copyrights\", \"characters\", and \"artist\" headers from the sidebar tag list."),
 			resize_link_style: newOption("dropdown", "full", "Resize Link Style", "Set how the resize links in the post sidebar options section will display. <tipdesc>Full:</tipdesc> Show the \"resize to window\", \"resize to window width\", and \"resize to window height\" links on separate lines. <tipdesc>Minimal:</tipdesc> Show the \"resize to window\" (W&H), \"resize to window width\" (W), and \"resize to window height\" (H) links on one line.", {txtOptions:["Full:full", "Minimal:minimal"]}),
 			search_add: newOption("dropdown", "disabled", "Search Add", "Modify the sidebar tag list by adding, removing, or replacing links in the sidebar tag list that modify the current search's tags. <tipdesc>Remove:</tipdesc> Remove any preexisting \"+\" and \"&ndash;\" links. <tipdesc>Link:</tipdesc> Add \"+\" and \"&ndash;\" links to modified versions of the current search that include or exclude their respective tags. <tipdesc>Toggle:</tipdesc> Add toggle links that modify the search box with their respective tags. Clicking a toggle link will switch between a tag being included (+), excluded (&ndash;), potentially included among other tags (~), and removed (&raquo;). Right clicking a toggle link will immediately remove its tag. If a tag already exists in the search box or gets entered/removed through alternative means, the toggle link will automatically update to reflect the tag's current status. <tiphead>Note</tiphead>The remove option is intended for users above the basic user level that want to remove the links. For users that can't normally see the links and do not wish to see them, this setting should be set to disabled.", {txtOptions:["Disabled:disabled", "Remove:remove", "Link:link", "Toggle:toggle"]}),
@@ -773,14 +790,14 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 							bbbNotice("Error retrieving post information. Your Danbooru API access is currently throttled. Please try again later.", -1);
 							bbbStatus("posts", "error");
 						}
-						else if (xmlhttp.status !== 0) {
-							if (xmlRetries < 1) {
+						else if (!bbb.flags.unloading) {
+							if (xmlhttp.status !== 0 && xmlRetries < 1) {
 								xmlRetries++;
 								fetchJSON(url, mode, optArg, xmlSession, xmlRetries);
 							}
 							else {
 								var linkId = uniqueIdNum(); // Create a unique ID.
-								var noticeMsg = bbbNotice('Error retrieving post information (JSON Code: ' + xmlhttp.status + ' ' + xmlhttp.statusText + '). (<a id="' + linkId + '" href="#">Retry</a>)', -1);
+								var noticeMsg = bbbNotice('Error retrieving post information (JSON Code: ' + xmlhttp.status + ' ' + xmlhttp.statusText + '). ' + (xmlhttp.status === 0 ? 'The connection was unexpectedly cancelled or timed out. ' : '') + '(<a id="' + linkId + '" href="#">Retry</a>)', -1);
 
 								bbbStatus("posts", "error");
 
@@ -798,6 +815,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 				}
 			};
 			xmlhttp.open("GET", url, true);
+			xmlhttp.timeout = 120000;
 			xmlhttp.send(null);
 		}
 	}
@@ -928,7 +946,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 					newUrl = postInfo.file_img_src;
 				}
 
-				imgContainer.innerHTML = '<div id="note-container"></div> <div id="note-preview"></div> <img alt="' + postInfo.tag_string + '" data-fav-count="' + postInfo.fav_count + '" data-flags="' + postInfo.flags + '" data-has-active-children="' + postInfo.has_active_children + '" data-has-children="' + postInfo.has_children + '" data-large-height="' + postInfo.large_height + '" data-large-width="' + postInfo.large_width + '" data-original-height="' + postInfo.image_height + '" data-original-width="' + postInfo.image_width + '" data-rating="' + postInfo.rating + '" data-score="' + postInfo.score + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" height="' + newHeight + '" width="' + newWidth + '" id="image" src="' + newUrl + '" /> <img src="about:blank" height="1" width="1" id="bbb-loader" style="position: absolute; right: 0px; top: 0px; display: none;" />';
+				imgContainer.innerHTML = '<div id="note-container"></div> <div id="note-preview"></div> <img alt="' + postInfo.tag_string_desc + '" data-fav-count="' + postInfo.fav_count + '" data-flags="' + postInfo.flags + '" data-has-active-children="' + postInfo.has_active_children + '" data-has-children="' + postInfo.has_children + '" data-large-height="' + postInfo.large_height + '" data-large-width="' + postInfo.large_width + '" data-original-height="' + postInfo.image_height + '" data-original-width="' + postInfo.image_width + '" data-rating="' + postInfo.rating + '" data-score="' + postInfo.score + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" height="' + newHeight + '" width="' + newWidth + '" id="image" src="' + newUrl + '" /> <img src="about:blank" height="1" width="1" id="bbb-loader" style="position: absolute; right: 0px; top: 0px; display: none;" />';
 
 				bbb.el.bbbLoader = document.getElementById("bbb-loader");
 
@@ -1020,7 +1038,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 				// Create the new post.
 				var childSpan = document.createElement("span");
 
-				childSpan.innerHTML = '<div id="post_' + postInfo.id + '" class="post post-preview' + postInfo.thumb_class + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" data-rating="' + postInfo.rating + '" data-flags="' + postInfo.flags + '" data-score="' + postInfo.score + '" data-parent-id="' + postInfo.parent_id + '" data-has-children="' + postInfo.has_children + '" data-id="' + postInfo.id + '" data-has-sound="' + postInfo.has_sound + '" data-width="' + postInfo.image_width + '" data-height="' + postInfo.image_height + '" data-approver-id="' + postInfo.approver_id + '" data-fav-count="' + postInfo.fav_count + '" data-pixiv-id="' + postInfo.pixiv_id + '" data-md5="' + postInfo.md5 + '" data-file-ext="' + postInfo.file_ext + '" data-file-url="' + postInfo.file_url + '" data-large-file-url="' + postInfo.large_file_url + '" data-preview-file-url="' + postInfo.preview_file_url + '"> <div class="preview"> <a href="/posts/' + postInfo.id + '"> <img alt="' + postInfo.md5 + '" src="' + postInfo.preview_img_src + '" /> </a> </div> <div class="comments-for-post" data-post-id="' + postInfo.id + '"> <div class="header"> <div class="row"> <span class="info"> <strong>Date</strong> <time datetime="' + postInfo.created_at + '" title="' + postInfo.created_at.replace(/(.+)T(.+)\..+-(.+)/, "$1 $2 -$3") + '">' + postInfo.created_at.replace(/(.+)T(.+):.+-.+/, "$1 $2") + '</time> </span> <span class="info"> <strong>Rating</strong> ' + postInfo.rating + ' </span> <span class="info"> <strong>Score</strong> <span> <span id="score-for-post-' + postInfo.id + '">' + postInfo.score + '</span> </span> </span> </div> <div class="row list-of-tags"> <strong>Tags</strong>' + tagLinks + '</div> </div> </div> <div class="clearfix"></div> </div>';
+				childSpan.innerHTML = '<div id="post_' + postInfo.id + '" class="post post-preview' + postInfo.thumb_class + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" data-rating="' + postInfo.rating + '" data-flags="' + postInfo.flags + '" data-score="' + postInfo.score + '" data-parent-id="' + postInfo.parent_id + '" data-has-children="' + postInfo.has_children + '" data-id="' + postInfo.id + '" data-has-sound="' + postInfo.has_sound + '" data-width="' + postInfo.image_width + '" data-height="' + postInfo.image_height + '" data-approver-id="' + postInfo.approver_id + '" data-fav-count="' + postInfo.fav_count + '" data-pixiv-id="' + postInfo.pixiv_id + '" data-md5="' + postInfo.md5 + '" data-file-ext="' + postInfo.file_ext + '" data-file-url="' + postInfo.file_url + '" data-large-file-url="' + postInfo.large_file_url + '" data-preview-file-url="' + postInfo.preview_file_url + '"> <div class="preview"> <a href="/posts/' + postInfo.id + '"> <img src="' + postInfo.preview_img_src + '" /> </a> </div> <div class="comments-for-post" data-post-id="' + postInfo.id + '"> <div class="header"> <div class="row"> <span class="info"> <strong>Date</strong> <time datetime="' + postInfo.created_at + '" title="' + postInfo.created_at.replace(/(.+)T(.+)\..+-(.+)/, "$1 $2 -$3") + '">' + postInfo.created_at.replace(/(.+)T(.+):.+-.+/, "$1 $2") + '</time> </span> <span class="info"> <strong>Rating</strong> ' + postInfo.rating + ' </span> <span class="info"> <strong>Score</strong> <span> <span id="score-for-post-' + postInfo.id + '">' + postInfo.score + '</span> </span> </span> </div> <div class="row list-of-tags"> <strong>Tags</strong>' + tagLinks + '</div> </div> </div> <div class="clearfix"></div> </div>';
 
 				// Prepare thumbnails.
 				prepThumbnails(childSpan);
@@ -1285,8 +1303,8 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 						else if (mode === "account_check")
 							accountUpdateHandler(docEl);
 					}
-					else if (xmlhttp.status !== 0) {
-						if (xmlRetries < 1) {
+					else if (!bbb.flags.unloading) {
+						if (xmlhttp.status !== 0 && xmlRetries < 1) {
 							xmlRetries++;
 							fetchPages(url, mode, optArg, xmlSession, xmlRetries);
 						}
@@ -1307,7 +1325,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 							else if (mode === "account_check")
 								msg = "Error checking account settings";
 
-							var noticeMsg = bbbNotice(msg + ' (HTML Code: ' + xmlhttp.status + ' ' + xmlhttp.statusText + '). (<a id="' + linkId + '" href="#">Retry</a>)', -1);
+							var noticeMsg = bbbNotice(msg + ' (HTML Code: ' + xmlhttp.status + ' ' + xmlhttp.statusText + '). ' + (xmlhttp.status === 0 ? 'The connection was unexpectedly cancelled or timed out. ' : '') + '(<a id="' + linkId + '" href="#">Retry</a>)', -1);
 
 							document.getElementById(linkId).addEventListener("click", function(event) {
 								if (event.button !== 0)
@@ -1322,6 +1340,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 				}
 			};
 			xmlhttp.open("GET", url, true);
+			xmlhttp.timeout = 120000;
 			xmlhttp.send(null);
 		}
 	}
@@ -2032,7 +2051,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		scrollDiv.appendChild(helpPage);
 
 		helpPage.bbbTextSection('Thumbnail Matching Rules', 'For creating thumbnail matching rules, please consult the following examples:<ul><li><b>tag1</b> - Match posts with tag1.</li><li><b>tag1 tag2</b> - Match posts with tag1 AND tag2.</li><li><b>-tag1</b> - Match posts without tag1.</li><li><b>tag1 -tag2</b> - Match posts with tag1 AND without tag2.</li><li><b>~tag1 ~tag2</b> - Match posts with tag1 OR tag2.</li><li><b>~tag1 ~-tag2</b> - Match posts with tag1 OR without tag2.</li><li><b>tag1 ~tag2 ~tag3</b> - Match posts with tag1 AND either tag2 OR tag3.</li></ul><br>Wildcards can be used with any of the above methods:<ul><li><b>~tag1* ~-*tag2</b> - Match posts with tags starting with tag1 or posts without tags ending with tag2.</li></ul><br>Multiple match rules can be specified by using commas or separate lines when possible:<ul><li><b>tag1 tag2, tag3 tag4</b> - Match posts with tag1 AND tag2 or posts with tag3 AND tag4.</li><li><b>tag1 ~tag2 ~tag3, tag4</b> - Match posts with tag1 AND either tag2 OR tag3 or posts with tag4.</li></ul><br>Tags can be nested/grouped together by using parentheses that only have spaces or commas next to them:<ul><li><b>( ~tag1 ~tag2 ) ( ~tag3 ~tag3 )</b> - Match posts with either tag1 OR tag2 AND either tag3 OR tag4.</li><li><b>tag1 ( tag2, tag3 tag4 )</b> - Match posts with tag1 AND tag2 or posts with tag1 AND tag3 AND tag4.</li><li><b>tag1 -( tag2 tag3 )</b> - Match posts with tag1 AND without tag2 AND tag3.</li><li><b>tag1 ~tag2 ~( tag3 tag4 )</b> - Match posts with tag1 and either tag2 OR tag3 AND tag4.</li></ul><br>The following metatags are supported:<ul><li><b>rating:safe</b> - Match posts rated safe. Accepted values include safe, explicit, and questionable.</li><li><b>status:pending</b> - Match pending posts. Accepted values include active, pending, flagged, banned, and deleted. Note that flagged posts also count as active posts.</li><li><b>user:albert</b> - Match posts made by the user Albert. Note that this tag will only work if you have a <b>moderator</b> level account or higher.</li><li><b>userid:1</b> - Match posts made by the user with an ID number of 1.</li><li><b>taggerid:1</b> - Match posts mostly tagged by the user with an ID number of 1.</li><li><b>approverid:1</b> - Match posts approved by the user with an ID number of 1. Accepted values include numbers, "any" for all posts with an approver, and "none" for posts without an approver.</li><li><b>source:http://www.4chan.org/</b> - Match posts with a source starting with http://www.4chan.org/. Accepted values include "any" for all posts with sources, "none" for all posts without sources, wildcard searches such as "*pixiv.net*" for posts with sources that contain pixiv.net, and non-wildcard searches that start matching at the beginning of a source.</li><li><b>isfav:true</b> - Match posts favorited under your current account. Accepted values include true and false.</li><li><b>group:hidden</b> or <b>g:hidden</b> - Match posts that match the tags in your group named \"hidden\".</li><li><b>pool:1</b> - Match posts that are in the pool with an ID number of 1. Accepted values include pool ID numbers, "series" for posts in series category pools, "collection" for posts in collection category pools, "any" for posts in any pool, "none" for posts not in a pool, "active" for posts in an active (not deleted) pool, and "inactive" for posts only in an inactive (deleted) pool.</li><li><b>parent:1</b> - Match posts that have the post with an ID number of 1 as a parent. Accepted values include post ID numbers, "any" for any posts with a parent, and "none" for posts without a parent.</li><li><b>child:any</b> - Match any posts that have children. Accepted values include "any" for any posts with children and "none" for posts without children.</li><li><b>id:1</b> - Match posts with an ID number of 1.</li><li><b>score:1</b> - Match posts with a score of 1.</li><li><b>favcount:1</b> - Match posts with a favorite count of 1.</li><li><b>height:1</b> - Match posts with a height of 1.</li><li><b>width:1</b> - Match posts with a width of 1.</li></ul><br>The id, score, favcount, width, and height metatags can also use number ranges for matching:<ul><li><b>score:&lt;5</b> - Match posts with a score less than 5.</li><li><b>score:&gt;5</b> - Match posts with a score greater than 5.</li><li><b>score:&lt;=5</b> or <b>score:..5</b> - Match posts with a score equal to OR less than 5.</li><li><b>score:&gt;=5</b> or <b>score:5..</b> - Match posts with a score equal to OR greater than 5.</li><li><b>score:1..5</b> - Match posts with a score equal to OR greater than 1 AND equal to OR less than 5.</li></ul>');
-		helpPage.bbbTextSection('Hotkeys', '<b>Posts</b><ul><li><b>B</b> - Open BBB menu.</li><li><b>1</b> - Resize to window.</li><li><b>2</b> - Resize to window width.</li><li><b>3</b> - Resize to window height.</li><li><b>4</b> - Reset/remove resizing.</li></ul><div style="font-size: smaller;">Note: Numbers refer to the main typing keypad and not the numeric keypad.</div><br><b>General</b><ul><li><b>B</b> - Open BBB menu.</li><li><b>E</b> - Toggle endless pages.</li><li><b>F</b> - Open quick search.</li><li><b>Shift + F</b> - Reset quick search.</li></ul>');
+		helpPage.bbbTextSection('Hotkeys', '<b>Posts</b><ul><li><b>B</b> - Open BBB menu.</li><li><b>1</b> - Resize to window.</li><li><b>2</b> - Resize to window width.</li><li><b>3</b> - Resize to window height.</li><li><b>4</b> - Reset/remove resizing.</li></ul><div style="font-size: smaller;">Note: Numbers refer to the main typing keypad and not the numeric keypad.</div><br><b>General</b><ul><li><b>B</b> - Open BBB menu.</li><li><b>E</b> - Toggle endless pages.</li><li><b>Shift + E</b> - Continue loading more pages after pausing during endless pages.</li><li><b>F</b> - Open quick search.</li><li><b>Shift + F</b> - Reset quick search.</li></ul>');
 		helpPage.bbbTextSection('Questions, Suggestions, or Bugs?', 'If you have any questions, please use the Greasy Fork feedback forums located <a target="_blank" href="https://greasyfork.org/scripts/3575-better-better-booru/feedback">here</a>. If you\'d like to report a bug or make a suggestion, please create an issue on GitHub <a target="_blank" href="https://github.com/pseudonymous/better-better-booru/issues">here</a>.');
 		helpPage.bbbTocSection();
 
@@ -3388,6 +3407,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 				case "8.0.1":
 				case "8.0.2":
 				case "8.1":
+				case "8.2":
 					break;
 			}
 
@@ -3887,13 +3907,12 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		if (optionsSection) {
 			var optionItems = optionsSection.getElementsByTagName("li");
 			var downloadRegex = /^\s*Download\s*$/i;
-			var downloadName = (getMeta("og:title") || "").replace(" - Danbooru", " - ");
 
 			for (i = 0, il = optionItems.length; i < il; i++) {
 				var optionItem = optionItems[i];
 
 				if (downloadRegex.test(optionItem.textContent)) {
-					optionItem.innerHTML = '<a download="' + downloadName + postInfo.md5 + '.' + postInfo.file_ext + '" href="' + postInfo.file_img_src + '?download=1">Download</a>';
+					optionItem.innerHTML = '<a download="' + postInfo.tag_string_desc + " - " + postInfo.md5 + '.' + postInfo.file_ext + '" href="' + postInfo.file_img_src + '?download=1">Download</a>';
 					break;
 				}
 			}
@@ -4706,13 +4725,18 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		// Cycle through each post and apply titles and borders.
 		for (i = 0, il = posts.length; i < il; i++) {
 			var post = posts[i];
+			var postSibling = post.nextSibling;
 			var img = post.getElementsByTagName("img")[0];
+
+			// Clean out text nodes between thumbnail articles.
+			if (postSibling && postSibling.nodeType === 3)
+				postSibling.parentNode.removeChild(postSibling);
 
 			if (!img)
 				continue;
 
 			var postInfo = post.bbbInfo();
-			var link = img.parentNode;
+			var link = img.bbbParent("A", 3);
 			var tagsStr = postInfo.tag_string || "";
 			var userStr = (postInfo.uploader_name ? " user:" + postInfo.uploader_name : "");
 			var ratingStr = (postInfo.rating ? " rating:" + postInfo.rating : "");
@@ -4812,7 +4836,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 	function createThumbHTML(postInfo, query) {
 		// Create a thumbnail HTML string.
-		return '<article class="post-preview' + postInfo.thumb_class + '" id="post_' + postInfo.id + '" data-id="' + postInfo.id + '" data-has-sound="' + postInfo.has_sound + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" data-rating="' + postInfo.rating + '" data-width="' + postInfo.image_width + '" data-height="' + postInfo.image_height + '" data-flags="' + postInfo.flags + '" data-parent-id="' + postInfo.parent_id + '" data-has-children="' + postInfo.has_children + '" data-score="' + postInfo.score + '" data-fav-count="' + postInfo.fav_count + '" data-approver-id="' + postInfo.approver_id + '" data-pixiv-id="' + postInfo.pixiv_id + '" data-md5="' + postInfo.md5 + '" data-file-ext="' + postInfo.file_ext + '" data-file-url="' + postInfo.file_url + '" data-large-file-url="' + postInfo.large_file_url + '" data-preview-file-url="' + postInfo.preview_file_url + '" data-source="' + postInfo.source + '" data-top-tagger="' + postInfo.keeper_data.uid + '" data-uploader-id="' + postInfo.uploader_id + '" data-normalized-source="' + postInfo.normalized_source + '" data-is-favorited="' + postInfo.is_favorited + '" data-file-url-desc="' + postInfo.file_url_desc + '"><a href="/posts/' + postInfo.id + query + '"><img src="' + postInfo.preview_img_src + '" alt="' + postInfo.tag_string + '"></a></article>';
+		return '<article class="post-preview' + postInfo.thumb_class + '" id="post_' + postInfo.id + '" data-id="' + postInfo.id + '" data-has-sound="' + postInfo.has_sound + '" data-tags="' + postInfo.tag_string + '" data-pools="' + postInfo.pool_string + '" data-uploader="' + postInfo.uploader_name + '" data-rating="' + postInfo.rating + '" data-width="' + postInfo.image_width + '" data-height="' + postInfo.image_height + '" data-flags="' + postInfo.flags + '" data-parent-id="' + postInfo.parent_id + '" data-has-children="' + postInfo.has_children + '" data-score="' + postInfo.score + '" data-fav-count="' + postInfo.fav_count + '" data-approver-id="' + postInfo.approver_id + '" data-pixiv-id="' + postInfo.pixiv_id + '" data-md5="' + postInfo.md5 + '" data-file-ext="' + postInfo.file_ext + '" data-file-url="' + postInfo.file_url + '" data-large-file-url="' + postInfo.large_file_url + '" data-preview-file-url="' + postInfo.preview_file_url + '" data-source="' + postInfo.source + '" data-top-tagger="' + postInfo.keeper_data.uid + '" data-uploader-id="' + postInfo.uploader_id + '" data-normalized-source="' + postInfo.normalized_source + '" data-is-favorited="' + postInfo.is_favorited + '" data-file-url-desc="' + postInfo.file_url_desc + '"><a href="/posts/' + postInfo.id + query + '"><picture><source media="(max-width: 660px)" srcset="' + postInfo.crop_img_src + '"><source media="(min-width: 660px)" srcset="' + postInfo.preview_img_src + '"><img src="' + postInfo.preview_img_src + '" alt="' + postInfo.tag_string + '"></picture></a></article>';
 	}
 
 	function createThumb(postInfo, query) {
@@ -5309,6 +5333,18 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		loadButtonDiv.id = "bbb-endless-load-div";
 		buttonDiv.appendChild(loadButtonDiv);
 
+		var loadToggle = function() {
+			// Continue page loading after hitting the pause interval.
+			if (bbb.endless.paused !== true)
+				return;
+
+			bbb.el.endlessLoadButton.style.display = "none";
+			bbb.el.endlessLoadButton.blur();
+			bbb.endless.paused = false;
+			bbb.endless.append_page = true;
+			endlessCheck();
+		};
+
 		var loadButton = bbb.el.endlessLoadButton = document.createElement("a");
 		loadButton.innerHTML = "Load More";
 		loadButton.href = "#";
@@ -5318,11 +5354,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			if (event.button !== 0)
 				return;
 
-			loadButton.style.display = "none";
-			loadButton.blur();
-			bbb.endless.paused = false;
-			bbb.endless.append_page = true;
-			endlessCheck();
+			loadToggle();
 			event.preventDefault();
 		}, false);
 		loadButtonDiv.appendChild(loadButton);
@@ -5348,8 +5380,9 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 		paginatorParent.insertBefore(buttonDiv, paginator);
 
-		// Create the hotkey.
+		// Create the hotkeys.
 		createHotkey("69", endlessToggle); // E
+		createHotkey("s69", loadToggle); // Shift + E
 
 		// Check the session default or original default value to see if endless pages should be enabled.
 		var sessionDefault = sessionStorage.getItem("bbb_endless_default");
@@ -6152,7 +6185,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 	function blacklistSmartView(el) {
 		// Set up the smart view event listeners.
 		var img = el.getElementsByTagName("img")[0];
-		var link = (img ? img.parentNode : undefined);
+		var link = (img ? img.bbbParent("A", 3) : undefined);
 
 		if (!link)
 			return;
@@ -6289,6 +6322,9 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 		// Cache any necessary info before leaving the page.
 		window.addEventListener("beforeunload", saveStateCache);
+
+		// Flag if the page is being unloaded in order to differentiate it from other potential XML interruptions.
+		window.addEventListener("beforeunload", function(){ bbb.flags.unloading = true; });
 	}
 
 	function formatInfo(postInfo) {
@@ -6321,6 +6357,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		postInfo.uploader_name = (isModLevel() ? postInfo.uploader_name : "");
 
 		// Custom BBB properties.
+		postInfo.tag_string_desc = tagStringDesc(postInfo);
 		postInfo.file_url_desc = postFileUrlDesc(postInfo);
 		postInfo.thumb_class = postClasses(postInfo);
 		postFileSrcs(postInfo);
@@ -6346,9 +6383,11 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		delete postInfo.has_sound;
 		delete postInfo.flags;
 		delete postInfo.thumb_class;
+		delete postInfo.crop_img_src;
 		delete postInfo.preview_img_src;
 		delete postInfo.file_img_src;
 		delete postInfo.large_file_img_src;
+		delete postInfo.tag_string_desc;
 		delete postInfo.normalized_source;
 
 		return postInfo;
@@ -6418,26 +6457,33 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 	function tagStringDesc(postInfo) {
 		// Create a Danbooru style post description.
-		if (typeof(postInfo.tag_string_artist) !== "string")
-			return "";
+		var downloadName = (getMeta("og:title") || "").replace(" - Danbooru", "");
+		var desc;
 
-		var artists = tagStringList(postInfo.tag_string_artist.replace(/(?:^|\s)banned_artist(?:$|\s)/, " "));
-		var characters = tagStringList(postInfo.tag_string_character.replace(/_\([^)]+\)(?:$|\s)/g, ""));
-		var copyrights = tagStringList(postInfo.tag_string_copyright);
+		if (gLoc === "post" && downloadName)
+			desc = downloadName;
+		else if (typeof(postInfo.tag_string_artist) === "string") {
+			var artists = tagStringList(postInfo.tag_string_artist.replace(/(?:^|\s)banned_artist(?:$|\s)/, " "));
+			var characters = tagStringList(postInfo.tag_string_character.replace(/_\([^)]+\)($|\s)/g, "$1"));
+			var copyrights = tagStringList(postInfo.tag_string_copyright);
 
-		if (characters !== "" && copyrights !== "")
-			copyrights = " (" + copyrights + ")";
+			if (characters !== "" && copyrights !== "")
+				copyrights = " (" + copyrights + ")";
 
-		if (artists !== "")
-			artists = " drawn by " + artists;
+			if (artists !== "")
+				artists = " drawn by " + artists;
 
-		return (characters + copyrights + artists).bbbSpaceClean() || "#" + postInfo.id;
+			desc = (characters + copyrights + artists).bbbSpaceClean() || "#" + postInfo.id;
+		}
+		else
+			desc = "";
+
+		return desc;
 	}
 
 	function tagStringFileUrlDesc(postInfo) {
 		// Create a Danbooru style post file/download description.
-		var desc = tagStringDesc(postInfo);
-		var fileDesc = (desc ? "__" + desc.replace(/[^0-9a-z]+/g, "_").replace(/^_+|_+$/g, "") + "__" : "");
+		var fileDesc = (postInfo.tag_string_desc ? "__" + postInfo.tag_string_desc.replace(/[^0-9a-z]+/g, "_").replace(/^_+|_+$/g, "") + "__" : "");
 
 		return fileDesc;
 	}
@@ -6457,6 +6503,14 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 	function postFileSrcs(postInfo) {
 		// Add the BBB file URL properties to post info.
 		postInfo.preview_img_src = postInfo.preview_file_url || bbbHiddenImg;
+
+		if (postInfo.md5) {
+			var pathChars = postInfo.md5.match(/^(..)(..)/);
+
+			postInfo.crop_img_src = "https://raikou3.donmai.us/crop/" + pathChars[1] + "/" + pathChars[2] + "/" + postInfo.md5 + ".jpg";
+		}
+		else
+			postInfo.crop_img_src = bbbHiddenImg;
 
 		if (disable_tagged_filenames && postInfo.file_url.indexOf("__") > -1) {
 			postInfo.file_img_src = postInfo.file_url.replace(postInfo.file_url_desc, "");
@@ -6613,7 +6667,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 			var thumb = document.createElement("a");
 			thumb.href = "/posts/" + postInfo.id;
-			thumb.innerHTML = '<img src="' + postInfo.preview_img_src + '" alt="' + postInfo.md5 + '">';
+			thumb.innerHTML = '<img src="' + postInfo.preview_img_src + '">';
 
 			if (before)
 				preview.insertBefore(thumb, before);
@@ -7926,6 +7980,8 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			'.bbb-quick-search-pinned #bbb-quick-search-pin, .bbb-quick-search-active.bbb-quick-search-pinned #bbb-quick-search-status {background-position: -144px -144px;}' + // Vertical pin.
 			'#bbb-quick-search-negate {background-position: -48px -129px;}' + // Negative sign.
 			'.bbb-quick-search-negated #bbb-quick-search-negate {background-position: -15px -192px;}' + // Negative sign in a dark circle.
+			'#bbb-quick-search.bbb-quick-search-negated {filter: invert(0.75);}' +
+			'#bbb-quick-search.bbb-quick-search-negated.bbb-quick-search-show {filter: invert(0);}' +
 			'#bbb-quick-search.bbb-quick-search-active {background-color: #DDDDDD;}' +
 			'#bbb-quick-search.bbb-quick-search-active.bbb-quick-search-show {background-color: #FFFFFF;}' +
 			'#bbb-quick-search-pin:focus, #bbb-quick-search-pin:hover, #bbb-quick-search-negate:focus, #bbb-quick-search-negate:hover {background-color: #CCCCCC;}' +
@@ -8109,24 +8165,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 		// Make the submit event search posts or reset the search.
 		searchForm.addEventListener("submit", function(event) {
-			var oldValue = bbb.quick_search.tags.bbbSpaceClean();
-			var oldNegate = bbb.quick_search.negated;
-			var curValue = searchInput.value.bbbSpaceClean();
-			var curNegate = searchDiv.bbbHasClass("bbb-quick-search-negated");
-
-			if (curValue === "" || (curValue === oldValue && curNegate === oldNegate))
-				quickSearchReset();
-			else {
-				bbb.quick_search.tags = searchInput.value;
-				bbb.quick_search.negated = searchDiv.bbbHasClass("bbb-quick-search-negated");
-
-				if (searchDiv.bbbHasClass("bbb-quick-search-pinned"))
-					sessionStorage.bbbSetItem("bbb_quick_search", JSON.stringify(bbb.quick_search));
-				else if (quick_search.indexOf("pinned") > -1)
-					quickSearchPinEnable();
-
-				quickSearchTest();
-			}
+			quickSearchToggle();
 
 			// Make autocomplete close without getting too tricky.
 			searchSubmit.focus();
@@ -8173,6 +8212,10 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			if (event.button === 0) {
 				if (event.shiftKey)
 					quickSearchPinToggle();
+				else if (event.ctrlKey) {
+					quickSearchNegateToggle();
+					quickSearchToggle();
+				}
 				else if (!searchDiv.bbbHasClass("bbb-quick-search-show"))
 					quickSearchOpen();
 				else
@@ -8235,6 +8278,30 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 		// Create the hotkeys.
 		createHotkey("70", quickSearchOpen); // F
 		createHotkey("s70", quickSearchReset); // SHIFT + F
+	}
+
+	function quickSearchToggle() {
+		// Submit the quick search or reset it if the current search is active.
+		var searchInput = bbb.el.quickSearchInput;
+		var searchDiv = bbb.el.quickSearchDiv;
+		var oldValue = bbb.quick_search.tags.bbbSpaceClean();
+		var oldNegate = bbb.quick_search.negated;
+		var curValue = searchInput.value.bbbSpaceClean();
+		var curNegate = searchDiv.bbbHasClass("bbb-quick-search-negated");
+
+		if (curValue === "" || (curValue === oldValue && curNegate === oldNegate))
+			quickSearchReset();
+		else {
+			bbb.quick_search.tags = searchInput.value;
+			bbb.quick_search.negated = searchDiv.bbbHasClass("bbb-quick-search-negated");
+
+			if (searchDiv.bbbHasClass("bbb-quick-search-pinned"))
+				sessionStorage.bbbSetItem("bbb_quick_search", JSON.stringify(bbb.quick_search));
+			else if (quick_search.indexOf("pinned") > -1)
+				quickSearchPinEnable();
+
+			quickSearchTest();
+		}
 	}
 
 	function quickSearchCheck() {
@@ -8459,8 +8526,8 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 			var targetTag = target.tagName;
 			var url; // If/else variable.
 
-			if (targetTag === "IMG" && target.parentNode)
-				url = target.parentNode.href;
+			if (targetTag === "IMG" && target.bbbParent("A", 3))
+				url = target.bbbParent("A", 3).href;
 			else if (targetTag === "A" && target.bbbHasClass("bbb-post-link", "bbb-thumb-link"))
 				url = target.href;
 
@@ -10211,7 +10278,7 @@ function bbbScript() { // Wrapper for injecting the script into the document.
 
 			$fields_multiple.autocomplete({
 				delay: 100,
-				autoFocus: true,
+				autoFocus: false,
 				focus: function() { return false; },
 				select: function(event, ui) {
 					var before_caret_text = this.value.substring(0, this.selectionStart);